% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foosball3.R
\name{create_foosball3_db}
\alias{create_foosball3_db}
\title{TODO}
\usage{
create_foosball3_db(file, ...)
}
\arguments{
\item{file}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
fl <- tempfile(fileext = ".sqlite")
create_foosball3_db(fl)

## Clean up:
unlink(fl, TRUE, TRUE)
}
