% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foosball3.R
\name{foosball3}
\alias{foosball3}
\title{Start the Foosball App}
\usage{
foosball3(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Renders the Graphical User Interface and starts the server
to interact with a Foosball SQLite database.
}
