% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{foosball3_create_db}
\alias{foosball3_create_db}
\title{TODO}
\usage{
foosball3_create_db(file, ...)
}
\arguments{
\item{file}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
fl <- tempfile(fileext = ".sqlite")
foosball3_create_db(fl)

## Clean up:
unlink(fl, TRUE, TRUE)
}
