% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{foosball3_export}
\alias{foosball3_export}
\title{Export an Foosball SQLite Database to csv Format}
\usage{
foosball3_export(source, destination, validate = FALSE, ...)
}
\arguments{
\item{source}{File path, where the source SQLite database can be found}

\item{destination}{File path, where the zipped collection of csv-files
should be stored}

\item{validate}{A \code{logical} value. If \code{TRUE}, the \code{source} will be imported
to check for validity (slower). If \code{FALSE}, \code{source} will be used directly without
checks.}

\item{...}{Ignored.}
}
\value{
\code{NULL} invisible
}
\description{
Export a database to a collection of csv-files. Note that you will
lose all views, relational information and constraints from the database.
}
\examples{
new_db <- tempfile()
foosball3_create_db(new_db)
new_export <- tempfile(fileext = ".zip")
foosball3_export(new_db, new_export)

## Clean up the files created in this example
unlink(new_db, TRUE, TRUE)
unlink(new_export, TRUE, TRUE)
}
