% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggests.R
\name{foosball3_list_suggests}
\alias{foosball3_list_suggests}
\title{List Packages Suggested by the \code{foosball3} Package}
\usage{
foosball3_list_suggests(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Returns a vector of package names that are suggested by
\code{foosball3} and required by its graphical user interface.
}
\description{
These are all packages required for the graphical user interface
(\code{\link[=foosball3]{foosball3()}}).
Make sure that they are installed
}
\examples{
foosball3_list_suggests()
}
