% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{foosball3_meta_data}
\alias{foosball3_meta_data}
\title{Get Meta Data for Tables in the Foosball Database Scheme}
\usage{
foosball3_meta_data(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
Returns a \code{data.frame} with description information
}
\description{
Returns a list of tables and fields in the Foosball database. In addition,
it contains a description of the table and each field.
}
\examples{
foosball3_meta_data()
}
