% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{foosball3_create_db}
\alias{foosball3_create_db}
\title{Create a New Foosball Database}
\usage{
foosball3_create_db(file, ...)
}
\arguments{
\item{file}{File path where the database needs to be stored}

\item{...}{Ignored}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Creates an SQLite file at the specified location. It will contain
all required relationships and constraints and some non-trivial data.
}
\examples{
fl <- tempfile(fileext = ".sqlite")
foosball3_create_db(fl)

## Clean up:
unlink(fl, TRUE, TRUE)
}
