% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{foosball3_generate_matches}
\alias{foosball3_generate_matches}
\title{TODO}
\usage{
foosball3_generate_matches(
  participants,
  tournament_type = "individual",
  tournament_phase = "qualification",
  options,
  progress,
  ...
)
}
\arguments{
\item{participants}{TODO}

\item{tournament_type}{TODO}

\item{tournament_phase}{TODO}

\item{options}{TODO}

\item{progress}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
participants <-
  dplyr::tibble(
    PARTICIPANT_ID = 5L:9L,
    QUALIFICATION_RATE = c(0.3, 0.5, 1, 1.3, 1.5)
  )
opts <- list(
  nsim = 100,
  revolutions = 2,
  weights = list(
    teamup         = 1,
    opposing       = 1,
    match_var      = 1,
    match_bal      = 1,
    match_bal_var  = 1,
    match_bal_extr = 1,
    part_bal       = 1,
    part_bal_var   = 1,
    part_bal_extr  = 1
  ),
  seed = 0
)

foosball3_generate_matches(
  participants, options = opts)
}
