% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{foosball3_generate_matches}
\alias{foosball3_generate_matches}
\title{Generate Matches for a Tournament Phase}
\usage{
foosball3_generate_matches(
  participants,
  tournament_type = "individual",
  tournament_phase = "qualification",
  options,
  progress,
  ...
)
}
\arguments{
\item{participants}{A \code{data.frame} with two columns: first \code{PARTICIPANT_ID},
a unique integer identifier for each participant. Second a column named
\code{QUALIFICATION_RATE}, it holds numeric values greater than zero. A qualification
rate of 0.5 indicates an average player. Values less than 0.5 are for less
experienced players. Values greater than 0.5 indicate more experienced players.}

\item{tournament_type}{Type of tournament for which to render matches.
At the moment only \code{"individual"} is supported. This is a tournament where
player play in variable team compositions, but for individual results.}

\item{tournament_phase}{The phase of the tournament. At the moment only
\code{qualification} is supported. These are matches preceding a semi finals, where
individuals can collect points individually to qualify for the semi finals.}

\item{options}{Options used for generating the matches. It can be used to
tweak the effort for balancing matches for player experience.}

\item{progress}{A callback function for reporting progress. It needs to be
a function that accepts 1 numeric argument. It is called by the generator,
with a number between 0 and 1 indicating the progress of the process.}

\item{...}{Ignored}
}
\value{
It returns a \code{data.frame} with three columns. Firstly, \code{MATCH_ID}
which is a unique integer identifier for each match. Secondly, \code{PARTICIPANT_ID},
which is from the input \code{participants}, and indicates the participants enrolled
in the indicated match. Thirdly, \code{POSITION_CODE}, indicating where the
participant is positioned for each specific match: \code{S1} = striker side 1,
\code{D1} = defender side 1, \code{S2} = striker side 2 and finally \code{D2} = defender side 2.
}
\description{
Generates matches for a specific phase in a foosball tournament for a
given set of participants.
}
\examples{
participants <-
  dplyr::tibble(
    PARTICIPANT_ID = 5L:9L,
    QUALIFICATION_RATE = c(0.3, 0.5, 1, 1.3, 1.5)
  )
opts <- list(
  nsim = 100,
  revolutions = 2,
  weights = list(
    teamup         = 1,
    opposing       = 1,
    match_var      = 1,
    match_bal      = 1,
    match_bal_var  = 1,
    match_bal_extr = 1,
    part_bal       = 1,
    part_bal_var   = 1,
    part_bal_extr  = 1
  ),
  seed = 0
)

foosball3_generate_matches(
  participants, options = opts)
}
