% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{foosball3_import_db}
\alias{foosball3_import_db}
\title{Import Foosball Tournament Data from a File}
\usage{
foosball3_import_db(file, target, ...)
}
\arguments{
\item{file}{An SQLite file or a zipped collection of csv files (created
with \code{\link[=foosball3_export_db]{foosball3_export_db()}}), to be imported into the standardised database
structure.}

\item{target}{Target file path where the clean database will be stored.
Existing files at this location may be overwritten.}

\item{...}{Ignored}
}
\value{
Creates a new clean copy of the database. Returns \code{NULL} invisibly.
}
\description{
If you have an older file with tournament results, you can try to
import those results with this function. It will create a new copy
of the database (using \code{\link[=foosball3_create_db]{foosball3_create_db()}}), making sure it
complies with the latest database scheme specification used by this package.
}
\examples{
tf <- tempfile()
export <- tempfile(fileext = ".zip")
import <- tempfile()
foosball3_create_db(tf)
foosball3_export_db(tf, export)
foosball3_import_db(export, import)

# Clean up example files
unlink(tf, TRUE, TRUE)
unlink(export, TRUE, TRUE)
unlink(import, TRUE, TRUE)
}
