% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{foosball3_import_db}
\alias{foosball3_import_db}
\title{TODO}
\usage{
foosball3_import_db(file, target, ...)
}
\arguments{
\item{file}{TODO}

\item{target}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
