% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggests.R
\name{foosball3_install_suggests}
\alias{foosball3_install_suggests}
\title{Install Packages Suggested by \code{foosball3}}
\usage{
foosball3_install_suggests(ignore = "mirai", update = TRUE, ...)
}
\arguments{
\item{ignore}{A vector of packages to ignore. By default
\code{"mirai"} is ignored, as this package is not needed for basic
functionality. It is recommended to install for better performance.}

\item{update}{A \code{logical} value. If \code{TRUE}, it will try to update
suggested packages that are already installed. If \code{FALSE}, it
will only install suggested packages that are not yet available.}

\item{...}{Arguments passed to \code{\link[utils:install.packages]{utils::install.packages()}}}
}
\value{
invisible \code{NULL}
}
\description{
Call this function if \code{\link[=foosball3_suggests_ok]{foosball3_suggests_ok()}} returns \code{FALSE}.
This will install all suggested packages required to run the graphical
user interface (\code{\link[=foosball3]{foosball3()}}).
}
\examples{
foosball3_list_suggests()
}
