% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggests.R
\name{foosball3_suggests_ok}
\alias{foosball3_suggests_ok}
\title{Check if All Suggested Packages Are Available}
\usage{
foosball3_suggests_ok(ignore = "mirai", ...)
}
\arguments{
\item{ignore}{A vector of packages to ignore. By default
\code{"mirai"} is ignored, as this package is not needed for basic
functionality. It is recommended to install for better performance.}

\item{...}{Arguments passed to \code{\link[=foosball3_list_suggests]{foosball3_list_suggests()}}}
}
\value{
A \code{logical} value, indicating of all required suggests are
available.
}
\description{
This function returns \code{TRUE} if all packages required for the
graphical user interface (GUI) (\code{\link[=foosball3]{foosball3()}}) are available.
}
\examples{
foosball3_list_suggests()
}
